requirements.txt
pandas~=1.3.1
Scrapy~=2.5.0
mysql-connector-python
beautifulsoup4